Pass args to another page, trigger demo and send email
Regex for validing email
